Added A-Javadoc
