+++ src/main/java/potaton/Potaton.java
@@ -4,6 +4,12 @@ import potaton.controller.Parser;
 import potaton.controller.UI;
 
 public class Potaton {
+
+    /**
+     * The chatbot itself, calls upon other classes to
+     * function.
+     * @param args
+     */
     public static void main(String[] args){
         UI.printWelcomeMessage();
         Parser.parseUserCommand();
