+++ src/main/java/potaton/controller/Storage.java
@@ -3,6 +3,7 @@ package potaton.controller;
 import potaton.task.Task;
 import potaton.task.ToDos;
 
+import java.io.FileNotFoundException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
@@ -10,9 +11,15 @@ import java.util.ArrayList;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.util.Scanner;
 
 public class Storage {
 
+    /**
+     * Check for existence of save file.
+     * If not found, create one.
+     * @param f
+     */
     public static void checkFileExistence(File f) {
         try {
             boolean fileExist = f.createNewFile();
@@ -28,6 +35,12 @@ public class Storage {
         }
     }
 
+    /**
+     * Loads and parse line into an Event object
+     * and add it to the task list
+     * @param line
+     * @param tasks
+     */
     public static void loadEvent(String line, ArrayList<Task> tasks) {
         String[] arg = line.split("] ", 2);
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMM d yyyy");
@@ -46,6 +59,12 @@ public class Storage {
         }
     }
 
+    /**
+     * Loads and parse line into a Deadline object
+     * and add it to the task list
+     * @param line
+     * @param tasks
+     */
     public static void loadDeadline(String line, ArrayList<Task> tasks) {
         String[] arg = line.split("] ", 2);
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMM d yyyy");
@@ -64,6 +83,12 @@ public class Storage {
         }
     }
 
+    /**
+     * Loads and parse line into a ToDos object
+     * and add it to the task list
+     * @param line
+     * @param tasks
+     */
     public static void loadToDo(String line, ArrayList<Task> tasks) {
         String[] arg = line.split("] ", 2);
         try {
@@ -82,13 +107,70 @@ public class Storage {
         }
     }
 
-    public static boolean checkIfDone(String arg) {
+    /**
+     * Check if given task is done by checking
+     * for presence of 'X' char in string.
+     * Returns true if found.
+     * @param arg
+     * @return
+     */
+    private static boolean checkIfDone(String arg) {
         if (arg.contains("X")){
             return true;
         }
         return false;
     }
 
+    /**
+     * Check for existence of save file and attempts
+     * to parse content from it if available. The
+     * parsed content is saved into the tasklist.
+     * If no save file found, a new save file is created.
+     * @param tasks
+     */
+    public static void loadFileContent(ArrayList<Task> tasks) {
+        File f = new File(UI.INIT_FILE_PATH);
+        Storage.checkFileExistence(f);
+        parseFileLine(tasks, f);
+    }
+
+    /**
+     * For each line in the save file, we attempt to
+     * parse it into one of the 3 task type object:
+     * Deadline, Event and Todo.
+     * @param tasks
+     * @param f
+     */
+    public static void parseFileLine(ArrayList<Task> tasks, File f) {
+        try {
+            Scanner s = new Scanner(f);
+            while (s.hasNext()) {
+                String line = s.nextLine();
+                String taskType = Character.toString(line.charAt(1));
+                switch (taskType) {
+                case UI.LOAD_TODO:
+                    Storage.loadToDo(line, tasks);
+                    break;
+                case UI.LOAD_DEADLINE:
+                    Storage.loadDeadline(line, tasks);
+                    break;
+                case UI.LOAD_EVENT:
+                    Storage.loadEvent(line, tasks);
+                    break;
+                default:
+                    UI.printText(UI.UNKNOWN_TASK);
+                    break;
+                }
+            }
+        } catch (FileNotFoundException e) {
+            UI.printText(UI.FILE_MISSING_MSG);
+        }
+    }
+
+    /**
+     * Saves task list contents into a save file.
+     * @param tasks
+     */
     public static void saveContentToFile(ArrayList<Task> tasks){
         try {
             FileWriter fw = new FileWriter(UI.INIT_FILE_PATH);
