+++ src/main/java/potaton/exception/PotatonException.java
@@ -1,4 +1,8 @@
 package potaton.exception;
 
+/**
+ * Public class of exception for self defined
+ * exception not found commonly
+ */
 public class PotatonException extends Exception{
 }
